requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
Pillow
google-genai
python-dotenv
